Tune MiniMax voice defaults

+++ services/minimax_tts.py
@@ -42,7 +42,7 @@ def tts_to_audio_bytes(text: str, audio_format: Optional[str] = None) -> Optiona
         "stream": False,
         "voice_setting": {
             "voice_id": MINIMAX_VOICE_ID,
-            "speed": int(MINIMAX_VOICE_SPEED),
+            "speed": float(MINIMAX_VOICE_SPEED),
             "vol": int(MINIMAX_VOICE_VOL),
             "pitch": int(MINIMAX_VOICE_PITCH),
             "emotion": MINIMAX_VOICE_EMOTION,
